requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
openai
langchain
langchain_openai
python-dotenv
pydantic-settings
httpx
tenacity
redis
pytest